Final commmit

+++ C969 Project/DeleteCustomer.cs
@@ -135,19 +135,34 @@ namespace C969_Project
                     //Grab List & convert
                     var list = getCustList();
                     IDictionary<string, object> dictionary = list.ToDictionary(pair => pair.Key, pair => pair.Value);
-                    DBHelper.deleteCustomer(dictionary);
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine(ex);
-                }
-                finally
-                {
+                    //First we need to check if appointments exist
+                    bool appoint = DBHelper.appointExist(dictionary["customerId"].ToString());
+                    if (appoint == false)
+                    {
+                        DBHelper.deleteCustomer(dictionary);
+
+                    }
+                    else
+                    {
+                        DialogResult youSure2 = MessageBox.Show("Deleting this customer will remove all of there associated appointments, contiune?", "", MessageBoxButtons.YesNo);
+                        if (youSure2 == DialogResult.Yes)
+                        {
+                            DBHelper.noCustAppointments(dictionary["customerId"].ToString());
+                            DBHelper.deleteCustomer(dictionary);
+                        }
+                        else {
+                            return;
+                        }
+                    }
                     ClearButton_Click(null, new EventArgs());
                     MessageBox.Show("Customer Record Deleted");
 
                     this.Owner.Show();
                     this.Close();
+                }
+                catch (Exception ex)
+                {
+                    Console.WriteLine(ex);
                 }
 
             }
